feat: expose text rendering via `textLayouter` service
Turn what was previously the `util/Text` utility into a service that
can be swapped out by users.

Keep backwards compatibility - it remains exposed as a utility for
utility use.

Related to #1023
