+++ lib/draw/index.js
@@ -1,5 +1,6 @@
 import DefaultRenderer from './DefaultRenderer';
 import Styles from './Styles';
+import TextLayouter from './TextLayouter';
 
 /**
  * @type { import('didi').ModuleDeclaration }
@@ -7,5 +8,6 @@ import Styles from './Styles';
 export default {
   __init__: [ 'defaultRenderer' ],
   defaultRenderer: [ 'type', DefaultRenderer ],
-  styles: [ 'type', Styles ]
+  styles: [ 'type', Styles ],
+  textLayouter: [ 'type', TextLayouter ]
 };
