+++ test/spec/draw/TextLayouterSpec.js
@@ -11,12 +11,12 @@ import {
   create as svgCreate
 } from 'tiny-svg';
 
-import TextUtil from 'lib/util/Text';
+import TextUtil from 'lib/draw/TextLayouter';
 
 import TestContainer from 'mocha-test-container-support';
 
 
-describe('util - Text', function() {
+describe('draw - TextLayouter', function() {
 
   var container;
